refactor: update `useSSRSafeId` to use base36 encoding for IDs
